Add shared Goss resources
As we currently do in circleci-images, we'll need to do a little bit of configuration and setup each time we want to run Goss tests on a particular image.

Namely, we'll need to append a few lines of code to the Dockerfile (those lines are contained in the `goss-add.Dockerfile` file), and then use a special entrypoint (`goss-entrypoint.sh`) for the version of the image that will be built and have Goss tests run against it.

As we will need to do this repeatedly, for every version of every image we build, it makes sense to store these resources in a common location.