Add run-goss-tests script
This file is adapted from the Bash function used in circleci-images. Rather than using convoluted logic to determine various filenames and paths, I've added variables to make the script more customizable.

Going forward, it may make more sense to turn this into an orb command in the `cci-dev/cimg` orb; but for now, too much is unclear to me about how variants will be regularly built/tested to want to orb-ify it yet. (For example, as soon as it's in an orb command, it becomes difficult to run in a loop, as compared to if it remains a shell script in this repository.)

The script is heavily commented, but generally:

- It is designed to be run once for each individual variant/tag that is being built, tested, and possibly published
- It expects a `goss.yaml` file to exist within the same directory from which `run-goss-tests.sh` is run
- It will create a new directory to serve as the build context for the Goss testing image, modify a given Dockerfile for the image to be tested, build the new testing image, run the tests in `goss.yaml`, and output results to a parameterized JUnit XML file for usage with CircleCI's Test Summary
- And finally, it cleans up after itself, `docker rm`'ing the Goss image after tests have finished